requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pypdf==3.17.4
PyPDF2==3.0.1
pytest==7.4.4
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
